+++ tests/config.lua
@@ -31,6 +31,7 @@ return function(t)
       auto_timezone = true,
       keymaps = false,
       time_bar = false,
+      time_bar_hover = false,
     })
 
     config.setup()
@@ -41,6 +42,7 @@ return function(t)
       auto_timezone = true,
       keymaps = false,
       time_bar = false,
+      time_bar_hover = false,
     })
   end)
 
@@ -133,6 +135,7 @@ return function(t)
       auto_timezone = true,
       keymaps = false,
       time_bar = false,
+      time_bar_hover = false,
     })
 
     config.setup()